moved resolvers to feature folder
