+++ src/app/app-routing.module.ts
@@ -28,14 +28,14 @@ import { AdminDashboardComponent } from "./admin-view/admin-dashboard/admin-dash
 import { DatasetFlowDetailsComponent } from "./dataset-flow/dataset-flow-details/dataset-flow-details.component";
 import { AccountComponent } from "./account/account.component";
 import { QueryExplainerComponent } from "./query-explainer/query-explainer.component";
-import { blockMetadataResolverFn } from "./common/resolvers/block-metadata.resolver";
-import { searchResolverFn } from "./common/resolvers/search.resolver";
-import { addPollingSourceResolverFn } from "./common/resolvers/add-polling-source.resolver";
-import { setTransformResolverFn } from "./common/resolvers/set-transform.resolver";
-import { addPushSourceResolverFn } from "./common/resolvers/add-push-source.resolver";
+import { blockMetadataResolverFn } from "./dataset-block/metadata-block/resolver/block-metadata.resolver";
+import { searchResolverFn } from "./search/resolver/search.resolver";
+import { addPollingSourceResolverFn } from "./dataset-view/additional-components/metadata-component/components/source-events/add-polling-source/resolver/add-polling-source.resolver";
+import { setTransformResolverFn } from "./dataset-view/additional-components/metadata-component/components/set-transform/resolver/set-transform.resolver";
+import { addPushSourceResolverFn } from "./dataset-view/additional-components/metadata-component/components/source-events/add-push-source/resolver/add-push-source.resolver";
 import { AccountSettingsTabs } from "./account/settings/account-settings.constants";
 import RoutingResolvers from "./common/resolvers/routing-resolvers";
-import { accountDatasetsResolverFn } from "./common/resolvers/account-datasets.resolver";
+import { accountDatasetsResolverFn } from "./account/additional-components/datasets-tab/resolver/account-datasets.resolver";
 import { AccountTabs } from "./account/account.constants";
 import { DatasetsTabComponent } from "./account/additional-components/datasets-tab/datasets-tab.component";
 import { AccountFlowsTabComponent } from "./account/additional-components/account-flows-tab/account-flows-tab.component";
@@ -44,47 +44,47 @@ import { FlowDetailsTabs } from "./dataset-flow/dataset-flow-details/dataset-flo
 import { FlowDetailsLogsTabComponent } from "./dataset-flow/dataset-flow-details/tabs/flow-details-logs-tab/flow-details-logs-tab.component";
 import { FlowDetailsUsageTabComponent } from "./dataset-flow/dataset-flow-details/tabs/flow-details-usage-tab/flow-details-usage-tab.component";
 import { FlowDetailsAdminTabComponent } from "./dataset-flow/dataset-flow-details/tabs/flow-details-admin-tab/flow-details-admin-tab.component";
-import { flowDetailsResolverFn } from "./common/resolvers/flow-details.resolver";
+import { flowDetailsResolverFn } from "./dataset-flow/dataset-flow-details/resolvers/flow-details.resolver";
 import { datasetInfoResolverFn } from "./common/resolvers/dataset-info.resolver";
 import { FlowDetailsSummaryTabComponent } from "./dataset-flow/dataset-flow-details/tabs/flow-details-summary-tab/flow-details-summary-tab.component";
-import { flowDetailsSummaryResolverFn } from "./common/resolvers/flow-details-summary.resolver";
+import { flowDetailsSummaryResolverFn } from "./dataset-flow/dataset-flow-details/tabs/flow-details-summary-tab/resolver/flow-details-summary.resolver";
 import { FlowDetailsHistoryTabComponent } from "./dataset-flow/dataset-flow-details/tabs/flow-details-history-tab/flow-details-history-tab.component";
 import { AccessTokensTabComponent } from "./account/settings/tabs/access-tokens-tab/access-tokens-tab.component";
 import { EmailsTabComponent } from "./account/settings/tabs/emails-tab/emails-tab.component";
-import { accountSettingsEmailResolverFn } from "./common/resolvers/account-settings-email.resolver";
-import { accountSettingsAccessTokensResolverFn } from "./common/resolvers/account-settings-access-tokens.resolver";
+import { accountSettingsEmailResolverFn } from "./account/settings/tabs/emails-tab/resolver/account-settings-email.resolver";
+import { accountSettingsAccessTokensResolverFn } from "./account/settings/tabs/access-tokens-tab/resolver/account-settings-access-tokens.resolver";
 import { DatasetViewTypeEnum } from "./dataset-view/dataset-view.interface";
 import { OverviewComponent } from "./dataset-view/additional-components/overview-component/overview.component";
-import { datasetOverviewTabResolverFn } from "./common/resolvers/dataset-view/dataset-overview-tab.resolver";
-import { datasetViewResolverFn } from "./common/resolvers/dataset-view/dataset-view.resolver";
+import { datasetOverviewTabResolverFn } from "./dataset-view/additional-components/overview-component/resolver/dataset-overview-tab.resolver";
+import { datasetViewResolverFn } from "./dataset-view/resolvers/dataset-view.resolver";
 import { FlowsComponent } from "./dataset-view/additional-components/flows-component/flows.component";
-import { datasetFlowsTabResolverFn } from "./common/resolvers/dataset-view/dataset-flows-tab.resolver";
+import { datasetFlowsTabResolverFn } from "./dataset-view/additional-components/flows-component/resolver/dataset-flows-tab.resolver";
 import { DataComponent } from "./dataset-view/additional-components/data-component/data.component";
 import { MetadataComponent } from "./dataset-view/additional-components/metadata-component/metadata.component";
 import { HistoryComponent } from "./dataset-view/additional-components/history-component/history.component";
 import { LineageComponent } from "./dataset-view/additional-components/lineage-component/lineage.component";
-import { datasetSettingsActiveSectionResolverFn } from "./common/resolvers/dataset-view/dataset-settings-active-section.resolver";
+import { datasetSettingsActiveSectionResolverFn } from "./dataset-view/additional-components/dataset-settings-component/resolvers/dataset-settings-active-section.resolver";
 import { DatasetSettingsComponent } from "./dataset-view/additional-components/dataset-settings-component/dataset-settings.component";
 import { DatasetSettingsGeneralTabComponent } from "./dataset-view/additional-components/dataset-settings-component/tabs/general/dataset-settings-general-tab.component";
 import { SettingsTabsEnum } from "./dataset-view/additional-components/dataset-settings-component/dataset-settings.model";
-import { datasetSettingsGeneralTabResolverFn } from "./common/resolvers/dataset-view/dataset-settings/dataset-settings-general-tab.resolver";
+import { datasetSettingsGeneralTabResolverFn } from "./dataset-view/additional-components/dataset-settings-component/tabs/general/resolver/dataset-settings-general-tab.resolver";
 import { DatasetSettingsTransformOptionsTabComponent } from "./dataset-view/additional-components/dataset-settings-component/tabs/transform-options/dataset-settings-transform-options-tab.component";
-import { datasetSettingsTransformTabResolverFn } from "./common/resolvers/dataset-view/dataset-settings/dataset-settings-transform-tab.resolver";
+import { datasetSettingsTransformTabResolverFn } from "./dataset-view/additional-components/dataset-settings-component/tabs/transform-options/resolver/dataset-settings-transform-tab.resolver";
 import { DatasetSettingsAccessTabComponent } from "./dataset-view/additional-components/dataset-settings-component/tabs/access/dataset-settings-access-tab/dataset-settings-access-tab.component";
 import { DatasetSettingsSchedulingTabComponent } from "./dataset-view/additional-components/dataset-settings-component/tabs/scheduling/dataset-settings-scheduling-tab.component";
 import { DatasetSettingsCompactingTabComponent } from "./dataset-view/additional-components/dataset-settings-component/tabs/compacting/dataset-settings-compacting-tab.component";
 import { DatasetSettingsSecretsManagerTabComponent } from "./dataset-view/additional-components/dataset-settings-component/tabs/variables-and-secrets/dataset-settings-secrets-manager-tab.component";
-import { datasetSettingsVarAndSecretsResolverFn } from "./common/resolvers/dataset-view/dataset-settings/dataset-settings-var-and-secrets.resolver";
-import { datasetSettingsSchedulingTabResolverFn } from "./common/resolvers/dataset-view/dataset-settings/dataset-settings-scheduling-tab.resolver";
-import { datasetSettingsAccessTabResolverFn } from "./common/resolvers/dataset-view/dataset-settings/dataset-settings-access-tab.resolver";
-import { datasetSettingsCompactionTabResolverFn } from "./common/resolvers/dataset-view/dataset-settings/dataset-settings-compaction-tab.resolver";
-import { datasetDataTabResolverFn } from "./common/resolvers/dataset-view/dataset-data-tab.resolver";
-import { datasetMetadataTabResolverFn } from "./common/resolvers/dataset-view/dataset-metadata-tab.resolver";
-import { datasetSettingsTabResolverFn } from "./common/resolvers/dataset-view/dataset-settings-tab.resolver";
-import { accountSettingsActiveTabResolverFn } from "./common/resolvers/account-settings-active-tab.resolver";
-import { accountActiveTabResolverFn } from "./common/resolvers/account-active-tab.resolver";
-import { datasetViewActiveTabResolverFn } from "./common/resolvers/dataset-view/dataset-view-active-tab.resolver";
-import { flowDetailsActiveTabResolverFn } from "./common/resolvers/flow-details-active-tab.resolver";
+import { datasetSettingsVarAndSecretsResolverFn } from "./dataset-view/additional-components/dataset-settings-component/tabs/variables-and-secrets/resolver/dataset-settings-var-and-secrets.resolver";
+import { datasetSettingsSchedulingTabResolverFn } from "./dataset-view/additional-components/dataset-settings-component/tabs/scheduling/resolver/dataset-settings-scheduling-tab.resolver";
+import { datasetSettingsAccessTabResolverFn } from "./dataset-view/additional-components/dataset-settings-component/tabs/access/dataset-settings-access-tab/resolver/dataset-settings-access-tab.resolver";
+import { datasetSettingsCompactionTabResolverFn } from "./dataset-view/additional-components/dataset-settings-component/tabs/compacting/resolver/dataset-settings-compaction-tab.resolver";
+import { datasetDataTabResolverFn } from "./dataset-view/additional-components/data-component/resolver/dataset-data-tab.resolver";
+import { datasetMetadataTabResolverFn } from "./dataset-view/additional-components/metadata-component/resolver/dataset-metadata-tab.resolver";
+import { datasetSettingsTabResolverFn } from "./dataset-view/additional-components/dataset-settings-component/resolvers/dataset-settings-tab.resolver";
+import { accountSettingsActiveTabResolverFn } from "./account/settings/resolver/account-settings-active-tab.resolver";
+import { accountActiveTabResolverFn } from "./account/resolver/account-active-tab.resolver";
+import { datasetViewActiveTabResolverFn } from "./dataset-view/resolvers/dataset-view-active-tab.resolver";
+import { flowDetailsActiveTabResolverFn } from "./dataset-flow/dataset-flow-details/resolvers/flow-details-active-tab.resolver";
 
 export const routes: Routes = [
     { path: "", redirectTo: ProjectLinks.DEFAULT_URL, pathMatch: "full" },
