+++ src/app/dataset-flow/dataset-flow-details/tabs/flow-details-summary-tab/resolver/flow-details-summary.resolver.ts
@@ -6,7 +6,7 @@
  */
 
 import { ActivatedRouteSnapshot, ResolveFn } from "@angular/router";
-import RoutingResolvers from "./routing-resolvers";
+import RoutingResolvers from "../../../../../common/resolvers/routing-resolvers";
 import { DatasetFlowByIdResponse } from "src/app/dataset-flow/dataset-flow-details/dataset-flow-details.types";
 
 export const flowDetailsSummaryResolverFn: ResolveFn<DatasetFlowByIdResponse> = (route: ActivatedRouteSnapshot) => {
