+++ src/app/dataset-view/additional-components/data-component/resolver/dataset-data-tab.resolver.ts
@@ -6,9 +6,9 @@
  */
 
 import { ResolveFn } from "@angular/router";
-import { datasetOverviewTabResolverFn } from "./dataset-overview-tab.resolver";
+import { datasetOverviewTabResolverFn } from "../../overview-component/resolver/dataset-overview-tab.resolver";
 import { DatasetOverviewTabData } from "src/app/dataset-view/dataset-view.interface";
 
-export const datasetMetadataTabResolverFn: ResolveFn<DatasetOverviewTabData> = (route, state) => {
+export const datasetDataTabResolverFn: ResolveFn<DatasetOverviewTabData> = (route, state) => {
     return datasetOverviewTabResolverFn(route, state);
 };
