+++ src/app/dataset-view/additional-components/dataset-settings-component/resolvers/dataset-settings-tab.resolver.ts
@@ -7,7 +7,7 @@
 
 import { ResolveFn } from "@angular/router";
 import { DatasetOverviewTabData } from "src/app/dataset-view/dataset-view.interface";
-import { datasetOverviewTabResolverFn } from "./dataset-overview-tab.resolver";
+import { datasetOverviewTabResolverFn } from "../../overview-component/resolver/dataset-overview-tab.resolver";
 
 export const datasetSettingsTabResolverFn: ResolveFn<DatasetOverviewTabData> = (route, state) => {
     return datasetOverviewTabResolverFn(route, state);
