+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/transform-options/resolver/dataset-settings-transform-tab.resolver.ts
@@ -7,9 +7,9 @@
 
 import { ActivatedRouteSnapshot, ResolveFn, RouterStateSnapshot } from "@angular/router";
 import { DatasetViewData } from "src/app/dataset-view/dataset-view.interface";
-import { datasetSettingsGeneralTabResolverFn } from "./dataset-settings-general-tab.resolver";
+import { datasetSettingsGeneralTabResolverFn } from "../../general/resolver/dataset-settings-general-tab.resolver";
 
-export const datasetSettingsSchedulingTabResolverFn: ResolveFn<DatasetViewData | null> = (
+export const datasetSettingsTransformTabResolverFn: ResolveFn<DatasetViewData | null> = (
     route: ActivatedRouteSnapshot,
     state: RouterStateSnapshot,
 ) => {
