+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/variables-and-secrets/resolver/dataset-settings-var-and-secrets.resolver.ts
@@ -7,7 +7,7 @@
 
 import { ResolveFn } from "@angular/router";
 import { DatasetViewData } from "src/app/dataset-view/dataset-view.interface";
-import { datasetSettingsGeneralTabResolverFn } from "./dataset-settings-general-tab.resolver";
+import { datasetSettingsGeneralTabResolverFn } from "../../general/resolver/dataset-settings-general-tab.resolver";
 
 export const datasetSettingsVarAndSecretsResolverFn: ResolveFn<DatasetViewData | null> = (route, state) => {
     return datasetSettingsGeneralTabResolverFn(route, state);
