+++ src/app/dataset-view/additional-components/flows-component/resolver/dataset-flows-tab.resolver.ts
@@ -6,7 +6,7 @@
  */
 
 import { ResolveFn } from "@angular/router";
-import { datasetOverviewTabResolverFn } from "./dataset-overview-tab.resolver";
+import { datasetOverviewTabResolverFn } from "../../overview-component/resolver/dataset-overview-tab.resolver";
 import { DatasetOverviewTabData } from "src/app/dataset-view/dataset-view.interface";
 
 export const datasetFlowsTabResolverFn: ResolveFn<DatasetOverviewTabData> = (route, state) => {
