+++ src/app/search/resolver/search.resolver.ts
